Folder Structure & Script Inclusion
I've set up the basic folder structure we've dealt with throughout our
homeworks, and updated the scripts (only touched up the Execute script
with correct file names) so that we can use them to make our lives
easier.

I also added an "Extra" folder to include the PDF and summer version of
the program, with the default output and log file so we can save these
with no needed hassle.